+++ src/main/java/com/tenut/asymmetrickeytool/PublicKey.java
@@ -17,38 +17,36 @@
  * limitations under the License.
  */
 
-package com.tenut.asynckeytool;
+package com.tenut.asymmetrickeytool;
 
 import java.security.KeyFactory;
 import java.security.KeyPair;
 import java.util.Base64;
 
-abstract class Key {
-  Key(KeyFactory factory, KeyPair keyPair) throws UnknownAsymmetricKeyAlgorithmException,
+abstract public class PublicKey extends Key {
+  PublicKey(KeyFactory factory, KeyPair keyPair) throws UnknownAsymmetricKeyAlgorithmException,
       InvalidAsymmetricKeyException {
-    newKey(factory, keyPair);
+    super(factory, keyPair);
   }
 
-  Key(KeyFactory factory, String encodedKey) throws InvalidEncodingException, InvalidAsymmetricKeyException,
+  PublicKey(KeyFactory factory, String encodedKey) throws InvalidAsymmetricKeyException,
       UnknownAsymmetricKeyAlgorithmException, InvalidEncodingException {
-    decode(factory, encodedKey);
+    super(factory, encodedKey);
   }
 
-  void decode(KeyFactory factory, String encoded) throws InvalidEncodingException, InvalidAsymmetricKeyException {
+  public String encrypt(String plainText) throws InvalidEncodingException {
+    return Base64.getEncoder().encodeToString(encryptData(plainText));
+  }
+
+  public boolean verify(String input, String output) throws InvalidEncodingException {
     try {
-      decodeKey(factory, Base64.getDecoder().decode(encoded));
+      return verifyData(input, Base64.getDecoder().decode(output));
     } catch (IllegalArgumentException e) {
-      throw new InvalidEncodingException("Key encoding not valid");
+      throw new InvalidEncodingException("Signature encoding not valid");
     }
   }
 
-  String asBase64String() throws InvalidAsymmetricKeyException {
-    return Base64.getEncoder().encodeToString(encodeKey());
-  }
+  abstract byte[] encryptData(String plainText) throws InvalidEncodingException;
 
-  abstract void newKey(KeyFactory factory, KeyPair keyPair) throws InvalidAsymmetricKeyException;
-
-  abstract void decodeKey(KeyFactory factory, byte[] encoded) throws InvalidAsymmetricKeyException;
-
-  abstract byte[] encodeKey();
+  abstract boolean verifyData(String input, byte[] output) throws InvalidEncodingException;
 }
